Created a course app and a class manager app to hold models, deleted the placeholder which had all but the user app. This is a proposed app structure for the project. User will need to be imported to the respective apps. Fields require validation and default values. Serializers will be needed once we are happy with the models.
