+++ backend/Class_manager_app/admin.py
