+++ backend/Class_manager_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ClassManagerAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'Class_manager_app'
