+++ backend/Class_manager_app/tests.py
