+++ backend/Class_manager_app/views.py
