+++ backend/Course_app/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class PlaceholderAppConfig(AppConfig):
+class CourseAppConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'placeholder_app'
+    name = 'Course_app'
